reorganize widgets
