+++ packages/jupyterlab-starters/src/widgets/async-component/index.tsx
