+++ packages/jupyterlab-starters/src/widgets/form/Theme/index.ts
@@ -1,3 +1,3 @@
 export { default } from './Theme';
-import '../../../style/form/index.css';
+import '../../../../style/form/index.css';
 export * from './Theme';
