+++ src/jupyter_starters/schema/v2.json
@@ -118,6 +118,10 @@
           "description": "JSON schema for a body to pass",
           "type": "object",
           "$ref": "#/definitions/json-schema"
+        },
+        "uiSchema": {
+          "description": "A react-jsonschema-form UI schema for customizing the selection of widgets https://react-jsonschema-form.readthedocs.io/en/latest/form-customization/#the-uischema-object",
+          "type": "object"
         }
       }
     },
